// Web.UI/NabinaERPV2.UI/src/app/models/setup-pages.model.ts
export class DesignationModel {
    id?: any;
    name?: string;
    sortOrder?: string;
    maxStaff?: string;
    isActive?: boolean;
}

export class DepartmentModel {
    id?: any;
    name?: string;
    shortName?: string;
    isActive?: boolean;
    sortOrder?: string;
}

export class SectionModel {
    id?: any;
    name?: string;
    sortOrder?: string;
    isActive?: boolean;
}